requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
future==1.0.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
MarkupSafe==3.0.3
moviepy==1.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.1
onnxruntime==1.24.3
opencv-python==4.11.0.86
packaging==26.0
pillow==12.1.1
piper-tts==1.4.1
proglog==0.1.12
protobuf==7.34.0
requests==2.32.5
scipy==1.16.3
setuptools==82.0.1
sympy==1.14.0
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
